Radar Competidores/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
python-pptx==0.6.23
PyYAML==6.0.2
streamlit
pyyaml
openai   # solo si después sí vas a integrar la API, por ahora no se usa
